+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Use an injected `init()` function to establish a postMessage communication channel
 - Inject styles from parent
+- Inject cookies from parent and allow iframe to postMessage with set cookies
 
 ## [0.0.1] - 2019-05-21